# Add benchmarking for qt vectors refresh