--- a/napari/_vendor/darkdetect/__init__.py
+++ b/napari/_vendor/darkdetect/__init__.py
@@ -4,7 +4,7 @@
 #  Distributed under the terms of the 3-clause BSD License.
 #-----------------------------------------------------------------------------
 
-__version__ = '0.7.0'
+__version__ = '0.7.1'
 
 import sys
 import platform