--- a/napari/_vendor/darkdetect/__main__.py
+++ b/napari/_vendor/darkdetect/__main__.py
@@ -1,0 +1,9 @@
+#-----------------------------------------------------------------------------
+#  Copyright (C) 2019 Alberto Sottile
+#
+#  Distributed under the terms of the 3-clause BSD License.
+#-----------------------------------------------------------------------------
+
+import darkdetect
+
+print('Current theme: {}'.format(darkdetect.theme()))