--- a/napari/_vendor/darkdetect/_mac_detect.py
+++ b/napari/_vendor/darkdetect/_mac_detect.py
@@ -6,7 +6,6 @@
 
 import ctypes
 import ctypes.util
-import typing
 
 try:
     # macOS Big Sur+ use "a built-in dynamic linker cache of all system-provided libraries"
@@ -70,5 +69,6 @@ def isDark():
 def isLight():
     return theme() == 'Light'
 
-def listener(callback: typing.Callable[[str], None]) -> None:
+#def listener(callback: typing.Callable[[str], None]) -> None:
+def listener(callback):
     raise NotImplementedError()