--- a/napari/layers/image/_tests/test_multiscale.py
+++ b/napari/layers/image/_tests/test_multiscale.py
@@ -234,10 +234,10 @@ def test_interpolation():
     with pytest.deprecated_call():
         assert layer.interpolation == 'bicubic'
 
-    layer.interpolation2d = 'bilinear'
+    layer.interpolation2d = 'linear'
     with pytest.deprecated_call():
-        assert layer.interpolation == 'bilinear'
-    assert layer.interpolation2d == 'bilinear'
+        assert layer.interpolation == 'linear'
+    assert layer.interpolation2d == 'linear'
 
 
 def test_colormaps():