… .github/workflows/deploy_docs.yml
… .github/workflows/docker-singularity-publish.yml
… docs/_static/version_switcher.json
… docs/guides/magicgui.md
… docs/howtos/layers/image.md
… docs/howtos/layers/labels.md
… docs/naps/3-spaces.md
… docs/plugins/find_and_install_plugin.md
… napari/benchmarks/benchmark_qt_viewer_vectors.py
… napari/layers/labels/_labels_key_bindings.py
… napari/layers/labels/labels.py
… napari/layers/points/_points_key_bindings.py
… napari/layers/shapes/_shapes_key_bindings.py
… napari/layers/shapes/shapes.py
… napari/layers/utils/_tests/test_layer_utils.py
… napari/layers/utils/text_manager.py
… napari/plugins/_tests/test_utils.py
… napari/plugins/hub.py
… napari/plugins/utils.py
… napari/qt/progress.py
… napari/settings/_application.py
… napari/utils/_injection/__init__.py
… napari/utils/_magicgui.py
… napari/utils/_proxies.py
… napari/utils/_tests/test_key_bindings.py
… napari/utils/_tests/test_translations.py
… napari/utils/events/containers/_selectable_list.py
… napari/utils/key_bindings.py
… napari/utils/settings/__init__.py
… napari/utils/theme.py