# Description
<!-- What does this pull request (PR) do? Why is it necessary? -->
<!-- Tell us about your new feature, improvement, or fix! -->
<!-- If your change includes user interface changes, please add an image, or an animation "An image is worth a thousand words!" -->
<!-- You can use https://www.cockos.com/licecap/ or similar to create animations -->
I've added a benchmark to check for regressions when switching to updating the mesh with every slice event. 

Based on my understanding, calling `refresh` should trigger both `on_data_change` _and_ `set_view_slice`. This means that `refresh` should encompass the mesh code in both locations and be as close to apples to apples as possible. 

That said, I'm not certain whether `refresh` should be called more than once in order to capture the additional burden of recalculating the mesh with every slice. I believe that in the old code, there was basically a check to see if anything had changed and if not, the mesh wasn't recalculated. This would mean that running refresh multiple times on the old code should be less computationally expensive than the new code _unless_ the reduction to only generate the mesh in the slice subset is sufficiently efficient. 

TLDR - I benchmarked against calling refresh in the test function once, and again with three consecutive calls to refresh.

**RESULTS:**


**call refresh once**
```
**old code**
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     2.81±0.02ms 
                 64     4.67±0.09ms 
                256      10.2±0.1ms 
                1024     36.3±0.8ms 
                4096     74.6±0.2ms 
               16384      110±1ms   
               65536      250±3ms   
              ======== =============


**new code**
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     2.93±0.03ms 
                 64     4.70±0.06ms 
                256      10.5±0.1ms 
                1024     38.1±0.9ms 
                4096     75.8±0.4ms 
               16384     122±0.8ms  
               65536      298±2ms   
              ======== =============
```

**call refresh 3 times**
```
**old code**
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16      8.30±0.1ms 
                 64     13.9±0.07ms 
                256      31.2±0.3ms 
                1024     105±0.6ms  
                4096      216±1ms   
               16384      326±6ms   
               65536      756±9ms   
              ======== =============

**new code**
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     8.63±0.02ms 
                 64     14.3±0.09ms 
                256      32.0±0.2ms 
                1024     106±0.7ms  
                4096      224±2ms   
               16384     358±0.8ms  
               65536      896±4ms   
              ======== =============
```   



## Type of change
<!-- Please delete options that are not relevant. -->
- [ ] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# References
<!-- What resources, documentation, and guides were used in the creation of this PR? -->
<!-- If this is a bug-fix or otherwise resolves an issue, reference it here with "closes #(issue)" -->

I added this benchmark as a qt benchmark because otherwise, the benchmark wasn't getting into `on_data_change`. It was strange because when I ran the same commands through pytest, I could have gotten away with not instantiating the viewer (I think). If there is a better approach, I'm happy to hear it! 


# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [ ] example: the test suite for my feature covers cases x, y, and z
- [ ] example: all tests pass with my change
- [ ] example: I check if my changes works with both PySide and PyQt backends
      as there are small differences between the two Qt bindings.  

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] If I included new strings, I have used `trans.` to make them localizable.
      For more information see our [translations guide](https://napari.org/developers/translations.html).