@kcpevey 2022-07-29
I think maybe I was supposed to [this](https://github.com/napari/napari/pull/4884) instead?